Move the methods that only work for pages from Collection to Page
addCollection(), getVersionToModify(), reindex(), getAreaCustomStyle(), outputCustomStyleHeaderItems() and cloneVersion() rely on the theme, the alias/master flags and the Page class: they have never worked for plain Collection instances, and the core only ever calls them on pages.

addCollection() had a check for Page instances, but when called on a plain Collection it would silently create a collection with no theme: it now belongs to Page too. The reindex() call of Collection::clearCollectionAttributes() is kept for pages only.
